PARQUET-1142: Extract non-Hadoop API from CodecFactory.
This adds CompressionCodecFactory to parquet-common. This addition
required moving org.apache.parquet.bytes into parquet-common from
parquet-encoding.

This also adds HadoopCodecs to parquet-hadoop, which implements the API.

Moving classes between artifacts required adding exclusions to semver
checks.
