+++ parquet-common/src/main/java/org/apache/parquet/bytes/ConcatenatingByteArrayCollector.java
